backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-multipart
Pillow
ultralytics
transformers
torch
datasets
bitsandbytes
peft
trl
sahi
